Switch to Double for all metrics
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC
Test: TrivialKotlinTest / TrivialStartupBenchmark # verified in Studio
Fixes: 197008210

Relnote: "Switch to floating point metrics throughout the benchmark
libraries (truncated in the Studio UI)."

Floating point metrics enable tracking smaller increments over time,
and avoid the hiding of instability which comes from rounding. This is
particularly important in faster benchmarks (very fast functions in
microbenchmark, or sub-ms frame timings in FrameTimingMetric in
Macrobenchmark (once we switch to perfetto-based frame timing).

Studio output is truncated to tenths of a ns / ms (micro/macrobench
respectively) for readability.

As part of this change, merged Stats/MetricResult as the point of the
wrapper is further reduced when all callers provide List<Double>.

+++ benchmark/benchmark-macro/src/main/java/androidx/benchmark/macro/perfetto/PerfettoResultsParser.kt
@@ -41,9 +41,9 @@ internal object PerfettoResultsParser {
         val durMs = getJSONObject("to_first_frame").getDouble("dur_ms")
         val fullyDrawnMs = optJSONObject("report_fully_drawn")?.getDouble("dur_ms")
 
-        val metricMap = mutableMapOf("startupMs" to durMs.toLong())
+        val metricMap = mutableMapOf("startupMs" to durMs)
         if (fullyDrawnMs != null) {
-            metricMap["fullyDrawnMs"] = fullyDrawnMs.toLong()
+            metricMap["fullyDrawnMs"] = fullyDrawnMs
         }
 
         val eventTimestamps = optJSONObject("event_timestamps")
